+++ provider/src/main/java/org/example/provider/Multiplication.java
@@ -8,7 +8,7 @@ public class Multiplication implements Calculate {
 
 
     @Override
-    public int result(int a, int b) {
+    public double result(double a, double b) {
         return a*b;
     }
 }
