+++ service/src/main/java/org/example/service/Calculate.java
@@ -1,5 +1,5 @@
 package org.example.service;
 
 public interface Calculate {
-    int result(int a, int b);
+    double result(double a, double b);
 }
